move CasInterval to integer_set
