+++ paddle/cinn/common/cas.cc
@@ -1770,86 +1770,6 @@ Expr ConvertCinnToCAS(Expr expr) {
   return expr;
 }
 
-/**
- * @brief Given an expr, visit it. If there is an ir::Min and its operands are 1
- * constant value and 1 inconstant value, return the constant min value. For
- * example, if a < min(5, b), then we get a < 5 and a < b. Using a < 5 to
- * simplify the condition ensures correctness, though not sufficient.
- */
-Expr ReplaceMinToConstant(Expr expr) {
-  Expr copied = ir::ir_utils::IRCopy(expr);
-  struct Mutator : public ir::IRMutator<ir::Expr*> {
-    void operator()(Expr* expr) { Visit(expr); }
-    void Visit(Expr* expr) { ir::IRMutator<>::Visit(expr, expr); }
-
-   private:
-    void Visit(const Min* op, Expr* expr) override {
-      auto a = op->a();
-      auto b = op->b();
-
-      Visit(&a);
-      Visit(&b);
-
-      auto min_a = op->a();
-      auto min_b = op->b();
-      if (min_a.is_constant() && !min_b.is_constant()) {
-        PADDLE_ENFORCE_EQ(
-            min_a->type().is_integer(),
-            true,
-            ::common::errors::InvalidArgument("Min a should be an integer."));
-        *expr = ir::ir_utils::IRCopy(min_a);
-      } else if (min_b.is_constant() && !min_a.is_constant()) {
-        PADDLE_ENFORCE_EQ(
-            min_b->type().is_integer(),
-            true,
-            ::common::errors::InvalidArgument("Min b should be an integer."));
-        *expr = ir::ir_utils::IRCopy(min_b);
-      }
-    }
-  };
-  Mutator()(&copied);
-  return copied;
-}
-
-/**
- * @brief Given an expr, visit it. If there is an ir::Max and its operands are 1
- * constant value and 1 inconstant value, return the constant max value.
- */
-Expr ReplaceMaxToConstant(Expr expr) {
-  Expr copied = ir::ir_utils::IRCopy(expr);
-  struct Mutator : public ir::IRMutator<ir::Expr*> {
-    void operator()(Expr* expr) { Visit(expr); }
-    void Visit(Expr* expr) { ir::IRMutator<>::Visit(expr, expr); }
-
-   private:
-    void Visit(const Max* op, Expr* expr) override {
-      auto a = op->a();
-      auto b = op->b();
-
-      Visit(&a);
-      Visit(&b);
-
-      auto max_a = op->a();
-      auto max_b = op->b();
-      if (max_a.is_constant() && !max_b.is_constant()) {
-        PADDLE_ENFORCE_EQ(
-            max_a->type().is_integer(),
-            true,
-            ::common::errors::InvalidArgument("Max a should be an integer."));
-        *expr = ir::ir_utils::IRCopy(max_a);
-      } else if (max_b.is_constant() && !max_a.is_constant()) {
-        PADDLE_ENFORCE_EQ(
-            max_b->type().is_integer(),
-            true,
-            ::common::errors::InvalidArgument("Max b should be an integer."));
-        *expr = ir::ir_utils::IRCopy(max_b);
-      }
-    }
-  };
-  Mutator()(&copied);
-  return copied;
-}
-
 Expr ConvertCasToCinn(Expr expr) {
   VLOG(7) << "Begin ConvertCasToCinn : " << expr;
