+++ paddle/cinn/common/integer_set.cc
@@ -689,5 +689,111 @@ ir::Expr EnhancedSimplifyModExpr(
   return copied;
 }
 
+CasInterval::CasInterval(ir::Expr expr_l, ir::Expr expr_r) {
+  VLOG(6) << "CasInterval is : [" << expr_l << ", " << expr_r << "].";
+  expr_r = ReplaceMinToConstant(expr_r);
+  expr_l = ReplaceMaxToConstant(expr_l);
+  expr_l = optim::ArithSimplify(expr_l);
+  expr_r = optim::ArithSimplify(expr_r);
+  VLOG(6) << "After simplify, CasInterval is : [" << expr_l << ", " << expr_r
+          << "].";
+
+  if (expr_l.is_constant() && expr_r.is_constant()) {
+    PADDLE_ENFORCE_EQ(
+        expr_l->type().is_integer(),
+        true,
+        ::common::errors::InvalidArgument("Expected expr_l to be an integer."));
+    PADDLE_ENFORCE_EQ(
+        expr_r->type().is_integer(),
+        true,
+        ::common::errors::InvalidArgument("Expected expr_r to be an integer."));
+    l = expr_l.as_int64();
+    r = expr_r.as_int64();
+    return;
+  }
+  e_l = expr_l;
+  e_r = expr_r;
+}
+
+/**
+ * @brief Given an expr, visit it. If there is an ir::Min and its operands are 1
+ * constant value and 1 inconstant value, return the constant min value. For
+ * example, if a < min(5, b), then we get a < 5 and a < b. Using a < 5 to
+ * simplify the condition ensures correctness, though not sufficient.
+ */
+ir::Expr CasInterval::ReplaceMinToConstant(ir::Expr expr) {
+  ir::Expr copied = ir::ir_utils::IRCopy(expr);
+  struct Mutator : public ir::IRMutator<ir::Expr*> {
+    void operator()(ir::Expr* expr) { Visit(expr); }
+    void Visit(ir::Expr* expr) { ir::IRMutator<>::Visit(expr, expr); }
+
+   private:
+    void Visit(const ir::Min* op, ir::Expr* expr) override {
+      auto a = op->a();
+      auto b = op->b();
+
+      Visit(&a);
+      Visit(&b);
+
+      auto min_a = op->a();
+      auto min_b = op->b();
+      if (min_a.is_constant() && !min_b.is_constant()) {
+        PADDLE_ENFORCE_EQ(
+            min_a->type().is_integer(),
+            true,
+            ::common::errors::InvalidArgument("Min a should be an integer."));
+        *expr = ir::ir_utils::IRCopy(min_a);
+      } else if (min_b.is_constant() && !min_a.is_constant()) {
+        PADDLE_ENFORCE_EQ(
+            min_b->type().is_integer(),
+            true,
+            ::common::errors::InvalidArgument("Min b should be an integer."));
+        *expr = ir::ir_utils::IRCopy(min_b);
+      }
+    }
+  };
+  Mutator()(&copied);
+  return copied;
+}
+
+/**
+ * @brief Given an expr, visit it. If there is an ir::Max and its operands are 1
+ * constant value and 1 inconstant value, return the constant max value.
+ */
+ir::Expr CasInterval::ReplaceMaxToConstant(ir::Expr expr) {
+  ir::Expr copied = ir::ir_utils::IRCopy(expr);
+  struct Mutator : public ir::IRMutator<ir::Expr*> {
+    void operator()(ir::Expr* expr) { Visit(expr); }
+    void Visit(ir::Expr* expr) { ir::IRMutator<>::Visit(expr, expr); }
+
+   private:
+    void Visit(const ir::Max* op, ir::Expr* expr) override {
+      auto a = op->a();
+      auto b = op->b();
+
+      Visit(&a);
+      Visit(&b);
+
+      auto max_a = op->a();
+      auto max_b = op->b();
+      if (max_a.is_constant() && !max_b.is_constant()) {
+        PADDLE_ENFORCE_EQ(
+            max_a->type().is_integer(),
+            true,
+            ::common::errors::InvalidArgument("Max a should be an integer."));
+        *expr = ir::ir_utils::IRCopy(max_a);
+      } else if (max_b.is_constant() && !max_a.is_constant()) {
+        PADDLE_ENFORCE_EQ(
+            max_b->type().is_integer(),
+            true,
+            ::common::errors::InvalidArgument("Max b should be an integer."));
+        *expr = ir::ir_utils::IRCopy(max_b);
+      }
+    }
+  };
+  Mutator()(&copied);
+  return copied;
+}
+
 }  // namespace common
 }  // namespace cinn
