+++ paddle/cinn/common/integer_set.h
@@ -14,11 +14,54 @@
 
 #pragma once
 #include <optional>
-#include "paddle/cinn/common/cas.h"
 #include "paddle/cinn/ir/ir.h"
+#include "paddle/cinn/optim/ir_simplify.h"
 
 namespace cinn {
 namespace common {
+
+/**
+ * Interval of a _Var_.
+ */
+struct CasInterval {
+  template <typename T>
+  CasInterval(T l, T r) : l(l), r(r) {
+    PADDLE_ENFORCE_LE(l,
+                      r,
+                      ::common::errors::InvalidArgument(
+                          "left should not be larger than right"));
+  }
+
+  /**
+   * @brief When iterator's upper_bound is an ir::Min of a constant value and a
+   * inconstant value, choose the constant value. When iterator's lower_bound is
+   * an ir::Max of a constant value and a inconstant value, choose the constant
+   * value. E.g: expr_l = max(x, 1) and expr_r = min(y,5): max(x, 1) <=
+   * iterator_i <= min(y,5)
+   *
+   * the bounds will be simplified to e_l = 1 and e_r = 5:
+   * 1 <= iterator_i <= 5
+   */
+  CasInterval(ir::Expr expr_l, ir::Expr expr_r);
+
+  ir::Expr ReplaceMinToConstant(ir::Expr expr);
+  ir::Expr ReplaceMaxToConstant(ir::Expr expr);
+
+  int l, r;
+  // Note: not verify l <= r and (e_l, e_r) has higher priority than (l, r)
+  ir::Expr e_l, e_r;
+
+  friend std::ostream& operator<<(std::ostream& os, const CasInterval& i) {
+    if (i.e_l.defined() && i.e_r.defined()) {
+      os << "Expr e_l Interval[" << i.e_l << ", " << i.e_r << "]";
+    } else {
+      os << "Int l Interval[" << i.l << ", " << i.r << "]";
+    }
+    return os;
+  }
+};
+
+using cas_intervals_t = absl::flat_hash_map<std::string, CasInterval>;
 
 cas_intervals_t CollectVarIntervalsOfExprs(const std::vector<ir::Expr>& exprs,
                                            bool is_lower_bound_zero = true);
